feat(security): Refactor SuperPaymaster validation checks
- Remove banned TIMESTAMP opcode usage by enforcing staleness via validUntil
- Implement validAfter-based rate limiting
- Fix griefing attack vector by updating timestamp in postOp unconditionally
- Update and pass 268 unit tests

+++ contracts/test/paymasters/superpaymaster/v3/SuperPaymasterRefundTest.t.sol
@@ -75,7 +75,7 @@ contract SuperPaymasterRefundTest is Test {
         vm.prank(entryPoint);
         (bytes memory context, uint256 validationData) = paymaster.validatePaymasterUserOp(userOp, opHash, maxCost);
         
-        assertEq(validationData, 0, "Validation should pass");
+        assertEq(uint160(validationData), 0, "Validation should pass");
         assertTrue(context.length > 0, "Context should be returned");
         
         // Check Operator Balance decreased by Max (logic: Pre-charge)
