+++ contracts/test/paymasters/superpaymaster/v3/SuperPaymasterV3.t.sol
@@ -354,7 +354,7 @@ contract SuperPaymasterTest is Test {
         (bytes memory context, uint256 validationData) = paymaster.validatePaymasterUserOp(op, opHash, 0.001 ether);
         vm.stopPrank();
 
-        assertEq(validationData, 0, "Validation should pass via Credit");
+        assertEq(uint160(validationData), 0, "Validation should pass via Credit");
         (address token, uint256 xAmount, address u, uint256 aAmount, bytes32 h, address opAddr) = abi.decode(context, (address, uint256, address, uint256, bytes32, address));
         assertEq(token, address(apnts));
         assertEq(opAddr, operator);
